Refactoring, removed usless code
